Converted Medical Tab of Patient to MVC
This makes it easier to keep track of.
